Remove warnings (#40)
* Support Dn for non-ESP32 boards

* remove note-c before re-add
* Squashed 'src/note-c/' content from commit 915db44

git-subtree-dir: src/note-c
git-subtree-split: 915db44ccfa27173e4003fa100fed84970755401

+++ src/note-c/n_helpers.c
@@ -1641,7 +1641,7 @@ uint32_t NoteMemAvailable()
     // Allocate progressively smaller and smaller chunks
     objHeader *lastObj = NULL;
     static long int maxsize = 35000;
-    for (long int i=maxsize; i>=sizeof(objHeader); i=i-sizeof(objHeader)) {
+    for (long int i=maxsize; i>=(long int)sizeof(objHeader); i=i-sizeof(objHeader)) {
         for (long int j=0;; j++) {
             objHeader *thisObj;
             thisObj = (objHeader *) _Malloc(i);
